Set default embedding precision to float32
- Update default SAVING_MODE from float16 to float32 in Generate_Embeddings.py

- Reorder SAVING_MODE dropdown options to default to float32 in EMAPSSN_Tools.py

- Update MCP Pipeline_Settings_Contracts.json schema default to float32

+++ src/EMAPSSN_Tools.py
@@ -1241,7 +1241,7 @@ class ToolsGUI(QMainWindow):
                         {
                             "var_name": "SAVING_MODE",
                             "type": "dropdown",
-                            "options": ["float16", "float32"],
+                            "options": ["float32", "float16"],
                             "display": "Saving Mode:"
                         },
                         {
